# upgrade: swap chalk→turbocolor; remove extraneous devDep