--- a/lib/helpers.js
+++ b/lib/helpers.js
@@ -5,7 +5,7 @@
 'use strict';
 
 const _ = require('lodash');
-const chalk = require('chalk')
+const tc = require('turbocolor')
 
 module.exports = {
   // Sets the constraints necessary during a `model.save` call.
@@ -72,7 +72,7 @@ module.exports = {
   },
 
   warn: function(msg) {
-    console.log(chalk.yellow(msg))
+    console.log(tc.yellow(msg))
   },
 
   deprecate: function(a, b) {