--- a/package.json
+++ b/package.json
@@ -25,10 +25,10 @@
   ],
   "dependencies": {
     "bluebird": "^3.4.3",
-    "chalk": "^2.3.0",
     "create-error": "~0.3.1",
     "inflection": "^1.5.1",
-    "lodash": "^4.17.10"
+    "lodash": "^4.17.10",
+    "turbocolor": "2.3.0"
   },
   "devDependencies": {
     "bookshelf-jsdoc-theme": "^0.2.0",
@@ -37,7 +37,6 @@
     "istanbul": "^0.4.5",
     "jsdoc": "^3.4.0",
     "knex": "^0.14.0",
-    "minimist": "^1.1.0",
     "mocha": "^5.0.0",
     "mysql": "^2.15.0",
     "pg": "^7.4.1",