Hi! @ricardograca! 👋 

## Summary

This PR replaces chalk for [turbocolor](https://github.com/jorgebucaran/turbocolor). Turbocolor is a ligther/faster drop-in replacement  for chalk (_10x_~_20x_ **faster** load time & runtime) with no dependencies. This is not a breaking change. 

## Motivation

Improve one dependency for a more performant one and remove an extraneous dependency.

## Proposed solution

Turbocolor has essentially the same API as chalk; it's just lighter (no deps) and faster. See full [benchmarks](https://github.com/jorgebucaran/turbocolor/tree/master/bench).

###### turbocolor/chalk

```
# Load Time
chalk: 15.190ms
turbocolor: 0.777ms

# All Colors
chalk × 8,729 ops/sec
turbocolor × 158,383 ops/sec

# Chained Colors
chalk × 1,838 ops/sec
turbocolor × 39,830 ops/sec

# Nested Colors
chalk × 4,049 ops/sec
turbocolor × 59,833 ops/sec
```

## Current PR Issues

None.

---

Let me know if this looks good to you and if you'd like to accept my contribution :)