@ricardograca 2018-07-18
Hey @jorgebucaran thanks for the contribution. It looks good to me. I was actually thinking of removing `chalk` altogether since it's only used to output a single color in a single place, so not sure how useful it is, but I'll accept this for now, especially because it has no dependencies.